html/cross-origin-opener-policy/multipart-popup.https.html

<!DOCTYPE html>
<html lang="en">
<title>Cross-Origin-Opener-Policy cannot be modified by multipart body parts</title>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script>
'use strict';

/**
 * These tests create a "parent" popup window which is an HTML document with a
 * specific Cross-Origin-Opener-Policy. The parent creates a "child" popup
 * window which is a non-HTML document with a Cross-Origin-Opener-Policy of its
 * own. The tests verify that when the Cross-Origin-Opener-Policy header is
 * among the response's headers, it us honored. The tests also verify that when
 * the same header is present among the multipart body part, it is *not*
 * honored.
 */
const make = ({t, parentCoop, resourceCoop, resourceBodyPartCoop, resourceName, validate}) => {
  const bc = new BroadcastChannel(token());
  bc.onmessage = t.step_func_done(({data}) => validate(data));
  const childLocation = `multipart.py?channel_name=${bc.name}` +
    (resourceCoop ? `&coop=${resourceCoop}` : '') +
    (resourceBodyPartCoop ? `&body_part_coop=${resourceBodyPartCoop}` : '');
  const parentLocation = 'resources/resource-popup.html' +
    `?channel_name=${bc.name}` +
    `&resource=${encodeURIComponent(childLocation)}` +
    `&resource_name=${resourceName}` +
    (parentCoop ? `&pipe=header(Cross-Origin-Opener-Policy,${parentCoop})` : '');

  open(parentLocation);

  t.add_cleanup(() => {
    bc.postMessage(null);
  });
};

async_test((t) => {
  make({
    t,
    parentCoop: '',
    resourceCoop: 'same-origin',
    resourceBodyPartCoop: '',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, null);
      assert_equals(data.closed, true);
    }
  });
}, 'parent COOP: unsafe-none; child COOP: same-origin; body part COOP: unsafe-none');

async_test((t) => {
  make({
    t,
    parentCoop: 'same-origin',
    resourceCoop: '',
    resourceBodyPartCoop: '',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, null);
      assert_equals(data.closed, true);
    }
  });
}, 'parent COOP: same-origin; child COOP: unsafe-none; body part COOP: unsafe-none');

async_test((t) => {
  make({
    t,
    parentCoop: 'same-origin',
    resourceCoop: 'same-origin',
    resourceBodyPartCoop: '',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, 'foobar');
      assert_equals(data.closed, false);
    }
  });
}, 'parent COOP: same-origin; child COOP: same-origin; body part COOP: unsafe-none');

async_test((t) => {
  make({
    t,
    parentCoop: '',
    resourceCoop: 'same-origin',
    resourceBodyPartCoop: 'same-origin',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, null);
      assert_equals(data.closed, true);
    }
  });
}, 'parent COOP: unsafe-none; child COOP: same-origin; body part COOP: same-origin');

async_test((t) => {
  make({
    t,
    parentCoop: 'same-origin',
    resourceCoop: '',
    resourceBodyPartCoop: 'same-origin',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, null);
      assert_equals(data.closed, true);
    }
  });
}, 'parent COOP: same-origin; child COOP: unsafe-none; body part COOP: same-origin');

async_test((t) => {
  make({
    t,
    parentCoop: 'same-origin',
    resourceCoop: 'same-origin',
    resourceBodyPartCoop: 'same-origin',
    resourceName: 'foobar',
    validate(data) {
      assert_equals(data.name, 'foobar');
      assert_equals(data.closed, false);
    }
  });
}, 'parent COOP: same-origin; child COOP: same-origin; body part COOP: same-origin');
</script>
</html>
